refactor(cron): split timer lifecycle
