+++ config/max-lines-baseline.txt
@@ -741,7 +741,6 @@ src/cron/service/jobs.ts
 src/cron/service/ops.test.ts
 src/cron/service/ops.ts
 src/cron/service/timer.regression.test.ts
-src/cron/service/timer.ts
 src/cron/store.test.ts
 src/daemon/launchd.test.ts
 src/daemon/launchd.ts
